split test folder to vrf, non-vrf and added new vrf.contract action test
