+++ .github/workflows/vrf.contracts.test.yaml
@@ -5,9 +5,8 @@ on:
     branches-ignore:
       - "master"
     paths:
-      - "contracts/**"
-      - "!contracts/**/VRFCoordinator.sol"
-      - "!contracts/**/VRFConsumerBase.sol"
+      - "contracts/**/VRFCoordinator.sol"
+      - "contracts/**/VRFConsumerBase.sol"
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -29,7 +28,7 @@ jobs:
         working-directory: ./contracts
       - run: yarn eslint
         working-directory: ./contracts
-      - run: npx hardhat test test/v0.1/non-vrf/*.cjs
+      - run: npx hardhat test test/v0.1/vrf/*.cjs
         working-directory: ./contracts
       - run: yarn compile
         working-directory: ./contracts
