+++ contracts/test/v0.1/non-vrf/Aggregator.test.cjs
@@ -8,7 +8,7 @@ const {
   parseSetRequesterPermissionsTx,
   deployDataFeedConsumerMock
 } = require('./Aggregator.utils.cjs')
-const { createSigners } = require('./utils.cjs')
+const { createSigners } = require('../utils.cjs')
 
 async function changeOracles(aggregator, removeOracles, addOracles) {
   const currentOracles = await aggregator.getOracles()
