+++ contracts/test/v0.1/non-vrf/CoordinatorBase.test.cjs
@@ -1,17 +1,17 @@
 const { expect } = require('chai')
 const { ethers } = require('hardhat')
 const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers')
-const { deploy: deployVrfConsumerMock } = require('./VRFConsumerMock.utils.cjs')
+const { deploy: deployVrfConsumerMock } = require('../vrf/VRFConsumerMock.utils.cjs')
 const {
   deploy: deployVrfCoordinator,
   setupOracle: setupVrfOracle,
   parseRandomWordsRequestedTx
-} = require('./VRFCoordinator.utils.cjs')
+} = require('../vrf/VRFCoordinator.utils.cjs')
 const { createAccount, addConsumer, deposit } = require('./Prepayment.utils.cjs')
-const { vrfConfig } = require('./VRFCoordinator.config.cjs')
+const { vrfConfig } = require('../vrf/VRFCoordinator.config.cjs')
 const { parseRequestCanceled } = require('./CoordinatorBase.utils.cjs')
 const { deploy: deployPrepayment } = require('./Prepayment.utils.cjs')
-const { parseKlay, createSigners } = require('./utils.cjs')
+const { parseKlay, createSigners } = require('../utils.cjs')
 
 async function deploy() {
   const {
