+++ contracts/test/v0.1/non-vrf/CoordinatorRevertFulfillment.test.cjs
@@ -7,8 +7,8 @@ const {
   parseRandomWordsRequestedTx,
   fulfillRandomWords,
   parseRandomWordsFulfilledTx
-} = require('./VRFCoordinator.utils.cjs')
-const { parseKlay } = require('./utils.cjs')
+} = require('../vrf/VRFCoordinator.utils.cjs')
+const { parseKlay } = require('../utils.cjs')
 const {
   deploy: deployRrCoordinator,
   setupOracle: setupRequestResponseCoordinator,
@@ -16,9 +16,9 @@ const {
   parseDataRequestFulfilledTx
 } = require('./RequestResponseCoordinator.utils.cjs')
 const { deploy: deployPrepayment, createAccount, deposit } = require('./Prepayment.utils.cjs')
-const { vrfConfig } = require('./VRFCoordinator.config.cjs')
+const { vrfConfig } = require('../vrf/VRFCoordinator.config.cjs')
 const { requestResponseConfig } = require('./RequestResponse.config.cjs')
-const { getBalance, createSigners } = require('./utils.cjs')
+const { getBalance, createSigners } = require('../utils.cjs')
 const oraklVrf = import('@bisonai/orakl-vrf')
 
 async function deploy() {
