+++ contracts/test/v0.1/non-vrf/Median.test.cjs
@@ -1,7 +1,7 @@
 const { expect } = require('chai')
 const { ethers } = require('hardhat')
 const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers')
-const { median } = require('./utils.cjs')
+const { median } = require('../utils.cjs')
 
 function medianBN(arr) {
   const arrBN = arr.map((x) => ethers.BigNumber.from(x))
